+++ crates/ignore-files/src/parse.rs
@@ -1,3 +1,4 @@
+pub mod bzr;
 pub mod charclass;
 mod common;
 pub mod git;